Projet BUT_3/semestre 6/application/Frontend/requirements.txt
streamlit
requests
python-docx
pymupdf
reportlab
googletrans-py